Fixes resolution of relative paths when mockery is `require()`d from multiple different modules
Mockery was using `module.parent` to resolve paths relative to the module that executed `require('mockery')`.
`module.parent` is only set to the *first* module that require()s mockery.  After that, mockery is cached and will
never be re-executed.

This fix deletes mockery from the require cache so that it's re-executed every time it's required.
`require('mockery')` always returns an instance of mockery that's bound to the module that required it.